# packages/protocol/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Luna API",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/v1/attachments": {
      "post": {
        "tags": [],
        "operationId": "attachments_upload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
    "/v1/attachments/{id}/content": {
      "get": {
        "tags": [],
        "operationId": "attachment_content",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Original image bytes"
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
    "/v1/attachments/{id}/thumbnail": {
      "get": {
        "tags": [],
        "operationId": "attachment_thumbnail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JPEG thumbnail bytes"
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
    "/v1/bootstrap": {
      "get": {
        "tags": [],
        "operationId": "bootstrap",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bootstrap"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
    "/v1/conversations": {
      "get": {
        "tags": [],
        "operationId": "conversations_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationList"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      },
      "post": {
        "tags": [],
        "operationId": "conversations_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
    "/v1/conversations/{id}": {
      "get": {
        "tags": [],
        "operationId": "conversation_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      },
      "patch": {
        "tags": [],
        "operationId": "conversation_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
    "/v1/conversations/{id}/abort": {
      "post": {
        "tags": [],
        "operationId": "conversation_abort",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "The active Pi operation was interrupted"
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
    "/v1/conversations/{id}/archive": {
      "post": {
        "tags": [],
        "operationId": "conversation_archive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The conversation was archived"
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
    "/v1/conversations/{id}/messages": {
      "get": {
        "tags": [],
        "operationId": "messages_list",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "beforeOrdinal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationMessages"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      },
      "post": {
        "tags": [],
        "operationId": "messages_send",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendMessageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
...
    "/v1/pairing/exchange": {
      "post": {
        "tags": [],
        "operationId": "pairing_exchange",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingExchangeResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pairing/request": {
      "post": {
        "tags": [],
        "operationId": "pairing_request",
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingCodeRequestResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repositories/{id}/icon": {
      "get": {
        "tags": [],
        "operationId": "repository_icon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detected repository icon bytes"
          }
        },
        "security": [
          {
            "deviceToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": ["code", "message", "retryable"],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": ["string", "null"]
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
...
      "AttachmentResponse": {
        "type": "object",
        "required": ["attachment"],
        "properties": {
          "attachment": {
            "$ref": "#/components/schemas/Attachment"
          }
        }
      },
...
      "Bootstrap": {
        "type": "object",
        "required": ["protocolVersion", "cursor", "device", "conversations"],
        "properties": {
          "conversations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Conversation"
            }
          },
          "cursor": {
            "type": "integer",
            "format": "int64"
          },
          "device": {
            "$ref": "#/components/schemas/Device"
          },
          "protocolVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "CreateConversationRequest": {
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "deviceToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}